Add exclude_botocore as input

+++ entrypoint.sh
@@ -10,7 +10,9 @@ install_zip_dependencies(){
 	echo "Installing and zipping dependencies..."
 	mkdir python
 	pip install --target=python -r "${INPUT_REQUIREMENTS_TXT}"
-	rm -rf ./python/botocore*
+	if [ "${INPUT_EXCLUDE_BOTOCORE}" == "true" ]; then
+		rm -rf ./python/botocore*
+	fi
 	zip -r dependencies.zip ./python
 	zipsplit -n 50000000 dependencies.zip
 	rm dependencies.zip
